Add ptr+ and ptr- expressions
They will likely be removed in the future, since raw pointer
arithmetics probably is not what we want to have in the final
language.

The most important part here is implementation that calls the
CreateGEP IRBuilder function. We will later create more explicit use
cases for it, e.g. bound-checekd array acces, calling methods with
correct field of correct structure lying in storage and so on.